renaming folder and add reference file
