--- MainAnalysis/20250802_ChargedHadronNeNe/README.md
@@ -1,4 +1,0 @@
-### Unblinding code ###
-
-
-
